Testing and bug fixing

+++ Capstone_IssueSprintManagement/backend/app/services/project_service.py
@@ -1,3 +1,5 @@
+import math
+
 from bson import ObjectId
 from bson.errors import InvalidId
 
@@ -6,11 +8,13 @@ from app.exceptions.project_exceptions import (
     ProjectNotFoundException,
     MemberAlreadyAssignedException,
     MemberNotAssignedException,
+    ActiveSprintExistsException,
 )
 from app.exceptions.user_exceptions import (
     UserNotFoundException,
     AdminAccessRequiredException,
 )
+from app.repositories.sprint_repository import SprintRepository
 from app.repositories.project_repository import ProjectRepository
 from app.repositories.user_repository import UserRepository
 from app.models.project_model import ProjectModel
@@ -24,6 +28,7 @@ class ProjectService:
     def __init__(self, db):
         self.project_repository = ProjectRepository(db)
         self.user_repository = UserRepository(db)
+        self.sprint_repository = SprintRepository(db)
 
     def _get_object_id(self, object_id: str):
         """
@@ -72,11 +77,17 @@ class ProjectService:
         result = self.project_repository.create_project(project)
         return str(result.inserted_id)
 
-    def get_all_projects(self):
+    def get_all_projects(
+            self,
+            page: int,
+            limit: int,
+        ):
         """
-        Fetch a list of all projects.
+        Fetch projects with pagination.
         """
-        projects = self.project_repository.get_all_projects()
+        total = self.project_repository.count_projects()
+
+        projects = self.project_repository.get_all_projects(page=page,limit=limit)
         project_list = []
 
         for project in projects:
@@ -116,7 +127,17 @@ class ProjectService:
                 }
             )
 
-        return project_list
+        return {
+            "items": project_list,
+            "total": total,
+            "page": page,
+            "limit": limit,
+            "total_pages":(
+                math.ceil(total / limit)
+                if total > 0
+                else 1
+            ),
+        }
 
     def update_project(self, project_id: str, project_data):
         """
@@ -142,6 +163,13 @@ class ProjectService:
 
         if not project:
             raise ProjectNotFoundException()
+
+        active_sprint = (
+            self.sprint_repository.find_active_sprint_by_project(project_object_id)
+        )
+
+        if active_sprint:
+            raise ActiveSprintExistsException()
 
         self.project_repository.delete_project(project_object_id)
